requirements.txt
area==1.1.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
geojson==3.1.0
idna==3.7
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
progressbar2==4.4.2
pycoral==2.0.0
python-utils==3.8.2
requests==2.32.3
soupsieve==2.6
tflite-runtime==2.5.0.post1
typing_extensions==4.12.2
urllib3==2.2.2